+++ ADEPO/common/calib1.h
@@ -1,7 +1,7 @@
 #ifndef CALIB1_H
 #define CALIB1_H
 
-#include "point3f.h"
+#include "point3d.h"
 #include "iostream"
 
 #include <QString>
@@ -9,15 +9,15 @@
 class Calib1
 {
 public:
-    Calib1(QString bcam, QString tpsCalib, Point3f coordPivot, Point3f coordAxis, float ccdToPivot, float ccdRotation) :
+    Calib1(QString bcam, QString tpsCalib, Point3d coordPivot, Point3d coordAxis, float ccdToPivot, float ccdRotation) :
         mBCAM(bcam), mTpsCalib(tpsCalib), mCoordPivot(coordPivot), mCoordAxis(coordAxis), mCcdToPivot(ccdToPivot), mCcdRotation(ccdRotation) {};
     virtual ~Calib1() {};
 
     //setter et getter
     QString getBCAM() const { return mBCAM; }
     QString getTpsCalib() const { return mTpsCalib; }
-    Point3f getCoordPivot() const { return mCoordPivot; }
-    Point3f getCoordAxis() const { return mCoordAxis; }
+    Point3d getCoordPivot() const { return mCoordPivot; }
+    Point3d getCoordAxis() const { return mCoordAxis; }
     int getDirection() const { return mCoordAxis.z() > 0 ? 1 : -1; }
     float getCcdToPivot() const { return mCcdToPivot; }
     float getCcdRotation() const { return mCcdRotation; }
@@ -40,8 +40,8 @@ protected:
 private:
     QString mBCAM;
     QString mTpsCalib;
-    Point3f mCoordPivot;
-    Point3f mCoordAxis;
+    Point3d mCoordPivot;
+    Point3d mCoordAxis;
     float mCcdToPivot;
     float mCcdRotation;
 };
